Refactor polar plot out of personal library chart
